# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AJKMart Super App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get user orders
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── Wallet ─────────────────────────────────────────────────────────────────
...
  /vendor/profile:
    patch:
      operationId: updateVendorProfile
      tags: [vendor]
      summary: Update vendor profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVendorProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorProfile"
  /vendor/store:
    get:
      operationId: getVendorStore
      tags: [vendor]
      summary: Get vendor store settings
      responses:
        "200":
          description: Store settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorStore"
    patch:
      operationId: updateVendorStore
      tags: [vendor]
      summary: Update vendor store settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorStore"
      responses:
        "200":
          description: Store updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorStore"
...
  /vendor/orders:
    get:
      operationId: getVendorOrders
      tags: [vendor]
      summary: Get vendor orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: after
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Vendor orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorOrdersResponse"
  /vendor/orders/{id}/status:
    patch:
      operationId: updateVendorOrderStatus
      tags: [vendor]
      summary: Update order status as vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Order status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorOrder"
...
components:
  schemas:
...
    CartItem:
      type: object
      properties:
        productId:
          type: string
        name:
          type: string
        price:
          type: string
        quantity:
          type: integer
        image:
          type: string
      required: [productId, name, price, quantity]
...
    Order:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        vendorId:
          type: string
        assignedRiderId:
          type: string
        type:
          type: string
          enum: [mart, food, rides, pharmacy, parcel, van, school]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, picked_up, out_for_delivery, delivered, cancelled]
        total:
          type: string
        deliveryAddress:
          type: string
        customerLat:
          type: number
        customerLng:
          type: number
        paymentMethod:
          type: string
          enum: [cash, wallet]
        riderId:
          type: string
        estimatedTime:
          type: string
        createdAt:
          type: string
      required: [id, userId, type, items, status, total, paymentMethod, createdAt]
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
      required: [orders, total]
    CreateOrderRequest:
      type: object
      properties:
        userId:
          type: string
        type:
          type: string
          enum: [mart, food, pharmacy, parcel]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        deliveryAddress:
          type: string
        paymentMethod:
          type: string
          enum: [cash, wallet]
      required: [userId, type, items, paymentMethod]
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, preparing, out_for_delivery, delivered, cancelled]
        riderId:
          type: string
      required: [status]

    # ── Wallet schemas ────────────────────────────────────────────────────────
...
    VendorProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        businessName:
          type: string
        storeName:
          type: string
        businessType:
          type: string
        phone:
          type: string
        email:
          type: string
        avatar:
          type: string
        isOnline:
          type: boolean
        isApproved:
          type: boolean
        rating:
          type: number
        totalOrders:
          type: number
      required: [id, userId, isOnline, isApproved]
    UpdateVendorProfileRequest:
      type: object
      properties:
        businessName:
          type: string
        storeName:
          type: string
        phone:
          type: string
        email:
          type: string
        avatar:
          type: string
        quickReplies:
          type: array
          items:
            type: string
    VendorStore:
      type: object
      properties:
        minOrderAmount:
          type: number
        deliveryFee:
          type: number
        estimatedDeliveryTime:
          type: string
        acceptsOnlinePayment:
          type: boolean
        acceptsCash:
          type: boolean
        description:
          type: string
        address:
          type: string
        city:
          type: string
...
    VendorOrder:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        status:
          type: string
        total:
          type: number
        deliveryAddress:
          type: string
        paymentMethod:
          type: string
        note:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        createdAt:
          type: string
      required: [id, userId, items, status, total, paymentMethod, createdAt]
    VendorOrdersResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/VendorOrder"
        total:
          type: integer
      required: [orders, total]
...